backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.11.0
python-dotenv==1.1.1
openai==1.107.2
PyJWT==2.10.1
httpx==0.28.1
psycopg[binary]==3.2.10
psycopg_pool==3.2.6
python-multipart==0.0.20
tiktoken==0.9.0
pytest==8.4.2
eval_type_backport==0.2.2
